Moved all Normalization into Normalizer.cs and made affected classes partial.
